Move ProcessorEnvironment and factory to SDK. ProcessorEnvironment also inherits from V2.
