+++ src/Microsoft.Performance.SDK/Processing/ProcessorEnvironment.cs
@@ -2,13 +2,12 @@
 // Licensed under the MIT License.
 
 using System;
-using Microsoft.Performance.SDK.Processing;
 
-namespace Microsoft.Performance.Toolkit.Engine
+namespace Microsoft.Performance.SDK.Processing
 {
     /// <inheritdoc cref="IProcessorEnvironment"/>
     public abstract class ProcessorEnvironment
-        : IProcessorEnvironment
+        : IProcessorEnvironmentV2
     {
         /// <inheritdoc/>
         public abstract ILogger CreateLogger(Type processorType);
@@ -21,6 +20,19 @@ namespace Microsoft.Performance.Toolkit.Engine
         public virtual IDynamicTableBuilder RequestDynamicTableBuilder(TableDescriptor descriptor)
         {
             return null;
+        }
+
+        /// <inheritdoc/>
+        public virtual IProcessorTableDataSynchronizationFactory TableDataSynchronizerFactory
+        {
+            get
+            {
+                return null;
+            }
+
+            protected set
+            {
+            }
         }
     }
 }
