+++ src/Microsoft.Performance.SDK/Processing/ProcessorEnvironmentFactory.cs
@@ -2,10 +2,9 @@
 // Licensed under the MIT License.
 
 using System;
-using Microsoft.Performance.SDK.Processing;
 using Microsoft.Performance.SDK.Processing.DataSourceGrouping;
 
-namespace Microsoft.Performance.Toolkit.Engine
+namespace Microsoft.Performance.SDK.Processing
 {
     /// <summary>
     ///     This is used to generate processor environments.
